fix the naming of gym_roboy folder. Added the feasible goal generator. Some basic visualization of the goal need to be fixed
