+++ gym_roboy/__init__.py
@@ -2,5 +2,5 @@ from gym.envs.registration import register
 
 register(
     id='msj-control-v0',
-    entry_point='gym-roboy.envs:MsjEnv',
+    entry_point='gym_roboy.envs:MsjEnv',
 )
