+++ gym_roboy/envs/cardsflow_bridge_scripts/compilation/compile_bridge.sh
