+++ gym_roboy/envs/cardsflow_bridge_scripts/init_bridge.sh
