+++ gym_roboy/envs/cardsflow_bridge_scripts/launch_simulation.sh
