+++ gym_roboy/envs/cardsflow_bridge_scripts/tmux_multiple_terminals_example/multiple_terminals_example.sh
